Refactor build system
