+++ Cargo.toml
@@ -6,7 +6,7 @@ edition = "2018"
 
 [lib]
 crate-type = ["cdylib", "rlib"]
-name = "chipolata"
+name = "libchipolata"
 path = "src/lib.rs"
 
 [[bin]]
@@ -17,14 +17,14 @@ required-features = ["cli"]
 [features]
 cli = ["structopt", "minifb", "rodio"]
 
-
 [dependencies]
 rand = "0.8.1"
 # CLI
 structopt = { version = "0.3.13", optional = true }
 minifb = { version = "0.19.1", optional = true }
 rodio = { version = "0.13.0", optional = true }
-# WASM
+
+[target.'cfg(target_arch = "wasm32")'.dependencies]
 wasm-bindgen = { version = "0.2" }
 getrandom = { version = "0.2", features = ["js"] }
 web-sys = { version = "0.3", features = ["console"] }
